+++ extension.js
@@ -869,7 +869,7 @@ function activate(context) {
   for (const [k, v] of Object.entries(context.workspaceState.get("ibmiBridge.hashes", {})))
     lastUpload.set(k, v);
   out = vscode.window.createOutputChannel("IBM i Bridge");
-  log(L.active("v0.11.0"));
+  log(L.active("v0.11.1"));
 
   context.subscriptions.push(
     vscode.commands.registerCommand("bridgeForI.pull", async () => {